requirements.txt
flask
gunicorn
transformers
opencensus-ext-azure
